+++ TestCode/src/DepGraphOpt.java
@@ -4,145 +4,13 @@ import java.io.*;
 import java.util.*;
 import java.util.stream.*;
 
-/**
- * Created by Mangesh on 11/26/2016.
- */
 
-
-class DependencyGraph
-{
-
-    //TODO: avoid creating multiple copies of the graph.
-    Map<CellRegion, Set<CellRegion>> forwardMap; //Depends -> DependOn
-    Map<CellRegion, Set<CellRegion>> reverseMap; // DependsOn -> Depends
-    static int count;
-
-    public DependencyGraph()
-    {
-        forwardMap = new HashMap();
-        reverseMap = new HashMap();
-    }
-
-    public void put(CellRegion depends, CellRegion dependsOn)
-    {
-        Set<CellRegion> values = forwardMap.get(depends);
-        if (values==null)
-        {
-            values = new HashSet();
-            forwardMap.put(depends, values);
-        }
-        values.add(dependsOn);
-
-        Set<CellRegion> keys = reverseMap.get(dependsOn);
-        if (keys==null)
-        {
-            keys = new HashSet();
-            reverseMap.put(dependsOn, keys);
-        }
-        keys.add(depends);
-
-    }
-
-    public Set<CellRegion> getDependsOn(CellRegion depends)
-    {
-        // Need to use intersection
-        return forwardMap.get(depends);
-    }
-
-
-    public Set<CellRegion> getDepends(CellRegion dependsOn)
-    {
-        return reverseMap.get(dependsOn);
-    }
-
-    public Set<CellRegion> getDependsSet()
-    {
-        return Collections.unmodifiableSet(forwardMap.keySet());
-    }
-
-
-    public Set<CellRegion> getDependsOnSet()
-    {
-        return Collections.unmodifiableSet(reverseMap.keySet());
-
-    }
-
-    /* Return the bounded box */
-    public CellRegion mergeDependsOn(Set<CellRegion> dependsOnSet) {
-
-        Set<CellRegion> dependsSet = new HashSet<>();
-        CellRegion boundingBox=null;
-
-        for (CellRegion dependsOn:dependsOnSet) {
-            dependsSet.addAll(deleteDependsOn(dependsOn));
-            if (boundingBox==null)
-                boundingBox = dependsOn;
-            else
-                boundingBox = boundingBox.getBoundingBox(dependsOn);
-        }
-        for (CellRegion depends:dependsSet)
-            put(depends, boundingBox);
-        return boundingBox;
-
-    }
-
-    public CellRegion mergeTwoDependsOn(CellRegion region1, CellRegion region2) {
-
-        Set<CellRegion> toMerge = new HashSet<>();
-        toMerge.add(region1);
-        toMerge.add(region2);
-        return mergeDependsOn(toMerge);
-    }
-
-    /* Remove value and if the list is empty remove key */
-    private void removeValue(Map<CellRegion, Set<CellRegion>> map, CellRegion key, CellRegion value) {
-        Set<CellRegion> valueSet = map.get(key);
-        valueSet.remove(value);
-        if (valueSet.isEmpty())
-            map.remove(key);
-
-    }
-
-    public Set<CellRegion> deleteDependsOn(CellRegion dependsOn)
-    {
-        Set<CellRegion> dependsSet = reverseMap.remove(dependsOn);
-        dependsSet.stream().forEach(e -> removeValue(forwardMap, e, dependsOn));
-        return dependsSet;
-    }
-
-    public Set<CellRegion> deleteDepends(CellRegion depends)
-    {
-        Set<CellRegion> dependsOnSet = forwardMap.remove(depends);
-        dependsOnSet.stream().forEach(e -> removeValue(reverseMap, e, depends));
-        return dependsOnSet;
-    }
-
-
-    public DependencyGraph copy()
-    {
-        DependencyGraph newGraph = new DependencyGraph();
-        for(Map.Entry<CellRegion, Set<CellRegion>> forwardEntry
-                :forwardMap.entrySet())
-        {
-            for(CellRegion dependsOn:forwardEntry.getValue())
-            {
-                newGraph.put(forwardEntry.getKey(), dependsOn);
-            }
-        }
-        return newGraph;
-    }
-
-
-    public int dependsOnSize() {
-        return reverseMap.size();
-    }
-}
-
-public class DTOpt {
+public class DepGraphOpt {
     // Dependency graph
-    static DependencyGraph originalGraph;
+
 
     public static void main(String[] args) throws IOException {
+        DependencyGraph originalGraph;
         originalGraph = new DependencyGraph();
 
         FileInputStream inputStream = new FileInputStream("TestCode/SampleGraph.txt");
@@ -164,23 +32,30 @@ public class DTOpt {
 
 
         //traverse(dt, originalGraph);
-        getSpace(originalGraph).forEach(e -> System.out.println("Size " + e.dependsOnSize() + " FP " + FPRate(originalGraph, e)));
+
+
+        getAllCandidates(originalGraph)
+                .forEach(e -> System.out.println("Size " + e.dependsOnSize() + " FP " + FPRate(originalGraph, e)));
 
         int memoryBudget = 3;
-        DependencyGraph sol = getSpace(originalGraph)
+        DependencyGraph sol = getAllCandidates(originalGraph)
                 .filter(e -> e.dependsOnSize() <= memoryBudget)
                 .min(Comparator.comparingDouble(e -> FPRate(originalGraph, e))).get();
+
         System.out.println("Solution");
         System.out.println("Size " + sol.dependsOnSize() + " FP Rate " + FPRate(originalGraph, sol));
+
         for (CellRegion region : sol.getDependsSet())
             System.out.println(region.toString() + " " + sol.getDependsOn(region));
 
     }
 
-    static Stream<DependencyGraph> getSpace(DependencyGraph inputGraph) {
+    public static Stream<DependencyGraph> getAllCandidates(DependencyGraph inputGraph) {
         Iterator<DependencyGraph> dependencyGraphIterator = new Iterator<DependencyGraph>() {
 
+            // Are we getting the fist solution?
             boolean first;
+            // Pull the next sub solution
             boolean pullNextSubSolution;
             // Index within the sub solution
             int subIndex;
@@ -210,9 +85,8 @@ public class DTOpt {
 
                 if (removedDependsOn != null) {
                     removedDependsSet = partial.deleteDependsOn(removedDependsOn);
-                    subSet = getSpace(partial).iterator();
+                    subSet = getAllCandidates(partial).iterator();
                 }
-                //System.out.println("Removed subset " + subSet);
 
             }
 
@@ -231,7 +105,10 @@ public class DTOpt {
 
             @Override
             public DependencyGraph next() {
+                // not longer is the first
                 first = false;
+
+                // Base case
                 if (subSet == null)
                     return inputGraph;
 
@@ -239,7 +116,7 @@ public class DTOpt {
                     subSolution = subSet.next();
                     subSolutionDependsOnList = subSolution.getDependsOnSet()
                             .stream().collect(Collectors.toCollection(ArrayList::new));
-                    subIndex = -1;
+                    subIndex = -1; // First candidate, where regions are not combined.
                     pullNextSubSolution = false;
                 }
 
@@ -253,23 +130,17 @@ public class DTOpt {
                 if (subIndex == subSolutionDependsOnList.size())
                     pullNextSubSolution = true;
                 return solution;
-
-
             }
         };
-
 
         return StreamSupport.stream(
                 Spliterators.spliteratorUnknownSize(
                         dependencyGraphIterator, Spliterator.DISTINCT), false);
-
-
     }
 
 
-
-
-    static void traverse(DependencyGraph newGraph, DependencyGraph originalGraph) {
+    /* Simple traverse, results in duplicates */
+    void traverse(DependencyGraph newGraph, DependencyGraph originalGraph) {
 
         int count = 0;
         List<CellRegion> dependsOnList = new ArrayList<>(newGraph.getDependsOnSet());
@@ -299,8 +170,8 @@ public class DTOpt {
         }
     }
 
-    static double FPRate(DependencyGraph originalGraph,
-                  DependencyGraph newGraph)
+    public static double FPRate(DependencyGraph originalGraph,
+                                DependencyGraph newGraph)
     {
         // Sum up the FP rates for each of the formula
         int original_cells = 0;
